requirements.txt
daphne==4.1.2
Django==5.1.7
django-ninja==1.3.0
pre-commit==4.2.0
pydantic-settings==2.8.1
psycopg==3.2.6
django-cors-headers==3.14.0